+++ src/components/common/drawer/index.tsx
@@ -132,7 +132,6 @@ export default function MobileDrawer({
       background: bgColor,
       display: 'flex',
       flexDirection: 'column',
-      // height: 'calc(100dvh - 36px)',
       position: 'fixed',
       bottom: 0,
       left: 0,
